Move business logic out of result component
